pyproject.toml
[project]
name = "mcp-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.25",
    "langchain-mcp-adapters>=0.0.11",
    "langchain-openai>=0.3.16",
    "langfuse>=2.60.4",
    "langgraph>=0.4.3",
    "mcp[cli]>=1.7.1",
    "nest-asyncio>=1.6.0",
    "openai-agents>=0.0.14",
    "pydantic-ai[logfire]>=0.1.10",
    "python-dotenv>=1.1.0",
]
